// src/types/user.ts
export interface User {
  user_id: number;
  first_name: string;
  last_name: string;
  email: string;
  address_line1: string;
  address_line2: string;
  country: string;
  city: string;
  phone1: string;
  phone2: string;
  watsapp_no: string;
  company_id: number;
  type: string;
  date_created: Date;
  date_modified: Date;
  group: string;
}